The code now completes when a "simple" TM simulation is run, reporting summaries and final totals for steps and nonzero tape entries...

+++ FastSim/Define.h
@@ -16,7 +16,7 @@ using namespace boost;
 
 #include "Exception.h"
 
-typedef int INTEGER;
+typedef long long INTEGER;
 
 typedef int STATE;
 typedef int SYMBOL;
